chore: 更新 golangci-lint 配置
- 为腾讯云CLS SDK异步生产者客户端添加Close方法到排除列表
- 保持对标准库io和os包关闭方法的忽略配置

+++ pkg/logger/cls.go
@@ -100,7 +100,7 @@ func ShutdownLogger(timeout time.Duration) {
 		// Close the producer instance
 		producerMu.Lock()
 		if producerInstance != nil {
-			producerInstance.Close(int64(timeout.Milliseconds()))
+			producerInstance.Close(timeout.Milliseconds())
 		}
 		producerMu.Unlock()
